Store positions as double

+++ src/main/java/com/techjar/vivecraftforge/util/PlayerTracker.java
@@ -82,9 +82,9 @@ public class PlayerTracker {
 	}
 
 	public static PacketUberPacket getPlayerDataPacket(UUID uuid, VRPlayerData data) {
-		PacketHeadData headData = new PacketHeadData(data.seated, data.head.posX, data.head.posY, data.head.posZ, data.head.rotW, data.head.rotX, data.head.rotY, data.head.rotZ);
-		PacketController0Data controller0Data = new PacketController0Data(data.handsReversed, data.controller0.posX, data.controller0.posY, data.controller0.posZ, data.controller0.rotW, data.controller0.rotX, data.controller0.rotY, data.controller0.rotZ);
-		PacketController1Data controller1Data = new PacketController1Data(data.handsReversed, data.controller1.posX, data.controller1.posY, data.controller1.posZ, data.controller1.rotW, data.controller1.rotX, data.controller1.rotY, data.controller1.rotZ);
+		PacketHeadData headData = new PacketHeadData(data.seated, (float)data.head.posX, (float)data.head.posY, (float)data.head.posZ, data.head.rotW, data.head.rotX, data.head.rotY, data.head.rotZ);
+		PacketController0Data controller0Data = new PacketController0Data(data.handsReversed, (float)data.controller0.posX, (float)data.controller0.posY, (float)data.controller0.posZ, data.controller0.rotW, data.controller0.rotX, data.controller0.rotY, data.controller0.rotZ);
+		PacketController1Data controller1Data = new PacketController1Data(data.handsReversed, (float)data.controller1.posX, (float)data.controller1.posY, (float)data.controller1.posZ, data.controller1.rotW, data.controller1.rotX, data.controller1.rotY, data.controller1.rotZ);
 		return new PacketUberPacket(uuid, headData, controller0Data, controller1Data, data.worldScale, data.height);
 	}
